do not use long

+++ src/cli.h
@@ -2,8 +2,8 @@
 #include <cstdlib>
 
 struct Params {
-    long N = 1000; // number of documents
-    long Q = 100; // number of queries
+    int N = 1000; // number of documents
+    int Q = 100; // number of queries
 };
 
 void parseCommandLine(int argc, char* argv[], Params& params) {
